for #381 adding authorization for controller actions
added the pundit gem for authorization which makes it easy to organize, permissions by controller actions.

in context of calendar_events, all users can view the calendar and edit events, but only the evnt owner can delete an event.

rspec seems straightforward as well.  I addec a spec/policies folder to contain the testing files and added a few tests for the controller permissions.